[TOWNS-2981] Optimize go routine memory usage by putting multiple syncs in a single session on the sync runner (#3061)
### Description

The sync runner previously started a new go routine with a sync session
containing a single stream for each stream tracked by the stream
tracker, which is used by both the notification service and the app
registry service. This was an issue because the notification service has
been observed with high memory usage and we cannot scale vertically
anymore. This change hopefully allows us to avoid sharding the service
for now.

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

- Added a new sync runner that:
  - is configurable
  - combines multiple tracked streams into a single sync session, and
- rotates through the available remotes for a stream if any remote
becomes unavailable.
 
- The notification and app registry service tests pass using this sync
runner
- additionally we have a couple of new norace tests that validate
correctness.

### Checklist

- [X] Tests added where required
- [X] Documentation updated where applicable
- [X] Changes adhere to the repository's contribution guidelines