node: fix improper error handling in stream reconciliation (#3124)
If stream conciliation fails because non of the remotes could provide
all requested miniblocks `err` is `nil` because remotes might be able to
return a valid response, just not with all miniblocks.

This error was wrapped in a RiverError but being `nil` resulted in a
river error with code unknown. The reconciliation task is rescheduled on
particular error codes in which this unknown code is not one of. As a
result the stream reconciliation task isn't rescheduled.