Subgraph fix for primary key spaces (#3165)
### Description

Subgraph has been throwing RangeErrors in pg on createdDate writes. This
PR short circuits if the createdDate is malformed logging the error.
Logs have also been formatted to return json with a timestamp which
should improve our ability to debug. Utility functions in the api have
been moved to utils.