Add a backwards compatible snapshot has to persisted miniblocks (#3243)
this is purely for debug information, we want to know if a miniblock
header had a snapshot, and the way we do that moving forward is via the
snapshot hash. So if we delete the snapshot from the header, put in a
fake hash. I don’t want to compute a real hash because it will be slow
and there will be nothing to check it agianst.