Shared syncer set backfilling mechanism (#3162)
This PR adds a way to catch up a new client on a stream that’s already
being synced by others. When a user starts syncing and adds a stream
with a custom cookie, but that stream is already live in the shared
syncer set, the node will first send all the missed changes since that
cookie. Once it has sent those past updates, it will continue sending
new updates as they happen. To do this, the node makes a backfill
request to the stream syncer, which then delivers the stream delta to
the shared channel before forwarding it on to the client.