node: implement new strategy for choosing stream nodes (#3171)
Implements a new approach for picking nodes to assign a stream
to. It picks N candidates pseudo-randomly from the operational nodes and
selects from these candidates replication factor best nodes. With best
defined as nodes that has the least number of streams assigned to them.
It tries to select nodes from a unique set of operators for correct
distribution.

It introduces a new on-chain config setting (default value=1) that is
used to determine the number of candidates to select the repl factor
best nodes from (replication factor+extra_candidates).