Node-2-node auth (#2438)
This PR implements mTLS auth for node-to-node communication within the
towns-protocol project. The changes include the addition of middleware
for requiring client certificates, certificate verification logic, and
e2e tests to ensure the functionality of mTLS.

The certificate contains a sender node Ethereum address and a signature
of cert's public key by the node's Ethereum private key so the server
can verify the signature inside the certificate and check the signer
address in the nodes registry.