node: include last error in returned error when context in backoff timer (#3364)
Several operations need to be retried after failure.
`base#BackoffTracker` provides functionality to retry an operator N
times or with a timer backoff that periodically retries the operation
until a context expires/cancelled. When the context expires the returned
error doesn't contain the given last error that gives context to the
operation that must be tried. The result is that the caller can miss
valuable error context. This change adds the operation context if
available.