Move the decryption extensions back to the sdk package (#3405)
I hope this doesn’t break anyone, we only moved it here because the
encryption package was the first we released and we wanted it to contain
our encryption/decryption protocol - but having it away from the rest of
the sdk makes things unnecessarily hard and also it’s worthless in
isolation without the greater context of the streams, syncing, and
persistence. It’s also a huge bottleneck in the app right now and we
need to figure out something else.
I did not re-combine with the “ClientDecryptionExtensions” class - i can
do that later in a separate commit