Protocol and SDK updates to support bots, take 2 (#3421)
This PR is built on top of contract and web3 library changes to support
apps; it extends them by adding protocol and sdk client compatibility
for bots.

The client has been updated to allow bots to create user streams by
specifying an app address to confirm they are bots, since they will not
necessarily be a member of any space when initializing.

The protocol now evaluates bot membership and bot entitlement through a
separate logical path:
- bot membership is evaluated via an isAppInstalled call on the space
contract.
- bot entitlement is evaluated by isAppInstalled.
- A new permission has been added - "Owner". This is to evaluate if the
user is a space owner. In addition, two new chain auth kinds have been
introduced: a chain auth that determines if a user is a bot, and one
that determines if a user is not a bot. This is to support validating
user stream inceptions for bot users, which contain the app contract
address, and to ensure that bots do not create GDM or DM streams.

This PR is a continuation of [a previous
PR](https://github.com/towns-protocol/towns/pull/3272) pushed to a
different branch.

### Description

<!-- Provide a clear and concise description of your changes and their
purpose -->

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines

---------

Co-authored-by: Miguel Nascimento <miguelgomes13@live.com>