Remove app registry check on DM/GDM creation. (#3458)
### Description

<!-- Provide a clear and concise description of your changes and their
purpose -->

We would like to delay deploying the app registry contract to all
production environments until we confirm that the contract operates as
expected. Thus, we will disable the App Registry check on GDM and DM
creation in order to not require the contract in all environments.

This change should fix the nodes on alpha.

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines