Update for key exchange eligibility (#3446)
I’ve seen cases where a user joins a stream and requests keys in the
same block. In this case, the key request will be received and enqueued,
and possibly processed before the membership is confirmed so the client
will fail.
We want the client to continue, and it is safe to continue, because the
client responding to key requests will check isEntitled on the
blockchain before sending any sensitive key material.