Implement proper app registry forwarding of slash commands (#3570)
We forward slash command messages now, only to the bot for whom the
command was addressed. However, in the case where the bot is specified,
it does not want messages forwarded. We do not forward slash commands
addressed to that bot. If two or more bots have the same command, we
will only forward to the bot for whom the command was addressed.

Tags have been extended to describe slash commands and target specific
apps to execute them.

The notification service ignores all slash command messages.

### Description

<!-- Provide a clear and concise description of your changes and their
purpose -->

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines