feat: add StreamBackwardsReconciliationThreshold on-chain setting
Implements on-chain configuration for the backwards reconciliation threshold.
This setting determines when to use backwards vs forward reconciliation based
on how many miniblocks a stream is behind.

If a stream is behind by more than this threshold, backwards reconciliation
will be used. Otherwise, forward reconciliation is used.

Default value is set to 50 miniblocks.