Update getMembershipStatus to propogate errors up the call stack (#3653)
We were seeing an issue in the stress test where clients were
incorrectly being minted new memberships because the stress test
evaluated those clients as not having an existing membership. The reason
for that is that this method was silencing errors and returning a false
membership result. Let's propagate the error here so that no one
concludes a user is not a member when the call actually resulted in an
error.

### Description

<!-- Provide a clear and concise description of your changes and their
purpose -->

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines