potentially fix flaky test (#3768)
This test is flaky. Maybe I'm missing something but:
- after random timeouts, clients are racing to fulfill a solicitation.
in this case there are 3 clients (alice, bob, charlie) with keys syncing
the same stream. chuck is soliciting keys.
- sometimes clients will not have had time to pick up that someone else
fulfilled the solicitation at the same time, and will also send a
fulfillment
- we were using `fulfillment.userId` for checking senders, but that id
belongs to the key solicitor. we need to use the `senderUserId` for this
approach to work