BaseContractShim: add retries to read contract (#3524)
### Description

The `BaseContractShim` provides a `read` contract which is used for
reading data. This PR updates the read contract methods to bake in
retries.

I only applied the retries to shims configured for Base networks - tests
failed with shims configured on River - there were timeouts b/c
getStreams retried instead of failing immediately, and I'm not sure if
that would have real world consequence. So for now let's just retry Base
reads, which is where we're getting failures in prod anyway.

This contract is not used for writes, as there is a separate `write`
contract.

### Checklist

- [x] Tests added where required