Avoid sending empty solicitations - attempt 2
this partially reverts https://github.com/towns-protocol/towns/pull/3773
we don’t want to overbroadcast
that pr just sent a bunch of non-actionable data to everyone in the stream.
this pr now silently sends keys if
a) the user has a new device and b) we don’t have any of the keys they requested, but we have keys
this is meant to help bootstrap a new client
now that i think about it, it would probably help to randomize the array in this case so everyone sends you different session ids, that would be neat. i’ll do that in a followup