Fix notifications apps configuration (#3884)
### Description

Since viper does not support arrays configurations via env var, Apps
field cannot be initialized automatically.
As a workaround, we have added the TownsApp and SenditApp fields, and we
copy their values to the Apps field, so the rest of the code can stay
`apps` agnostic.
in the future, this configuration (both *App and Apps) will be moved to
a database, to allow dynamic config without code changes

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines