[TOWNS-26905] Fix app registry sql deadlocks by relaxing isolation level of select queries (#3900)
### Description

PR contains a change that allows an implementer of a PG store to
optionally override the default isolation level of any transaction, as
well as a change in the app registry to downgrade the isolation level of
all non-queue related transactions to read committed.

I initially experimented with a row locking approach to see if this
would resolve the deadlocking issue we were seeing in tests. However,
when I enabled more verbose and detailed PostgreSQL logs, I saw that the
deadlocks in tests were actually SSI serialization aborts occuring
during unique index checks while updating the device_key column in the
registry table. The device_key column is only updated during the
register webhook call, which is called in parallel at the beginning of
norace tests. Reducing the isolation level of the RegisterWebhook method
on the app registry store turned out to be the real fix.

### Changes

<!-- List the specific changes made in this PR, for example:
- Added/modified feature X
- Fixed bug in component Y
- Refactored module Z
- Updated documentation
-->

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines