Refactor web3 and towns environment builders (#4180)
i've left this in several commits so it's reviewable piece by piece, but
will squash everything once the time comes
some platforms (chrome webworkers) throw a fit if we access import.meta
and refuse to compile so we need to give the user an option to pass in
an environment. Passing it in every situation that needed it was
infeasible, i made multiple mistakes trying to make that refactor and
didn't catch everything.
this pr wraps the envs into sudo-builders that capture the optional
environment options and pass them along when necessary.