Added an index on spaces table on subgraph setup hook (#4374)
Subgraph has an optimization to create the indexes only after the historic indexing is done.
We are reading from the table while we index, and if this index is not added before indexing starts, then postgres will do a full table scan for every SpaceOwner:Transfer event, which make it run slower the more spaces we index.