Bot tests - don’t sync the “botClient” it affects other tests
initializeUser starts syncing (yeah this is my bad i guess), so if you do this in the test you always have a user client in the background doing key exchange and it throws off the tests. for a sec i thought the bot server was already doing key exchange because alice was getting keys when there was no one but the bot there to send them to her. whoops.