node: use genesis hash from SealEphemeralStream response (#4450)
This PR fixes a TODO that when an ephemeral stream was sealed the
genesis miniblock hash in the response wasn't used but instead the
optimistic approach was used to use the local genesis hash when enough
nodes sealed the stream successful.

With this PR the reported genesis hashes are counted and the stream is
considered to be successfully sealed when enough votes for the same
genesis miniblock where casted.