Forward interaction responses to bots when setting is Reply only
fixes TOWNS-31894
I spent kinda a lot of time trying to extend the go test for this, but it never worked and I couldn’t figure out why. I tried overloading the sendMessage helper so that i could send an interaction response, which caused me to need to update a bunch of other things, and then the message just never got to the event processer. idk.