Add Cleanup Routine in Persistence Store when stores 95% full (#4531)
It's been so long since we had a problem I kinda forgot about this. We
only save small amounts of info, but we never clean up the stores. This
is all local cache data, deleting it just causes us to re fetch it
Tested by checking the "Simulate custom stoarge quota" checkbox in
chrome and watching the memory go way back down to reasonable levels