Bot tests - clean up handlers after each test
they were interfeering with each other - for example, after the tip test we were tipping every single message.
We could also add a global reset and call that instead, or add a member variable and keep track of subscriptions in the bot object… but it seems only useful in tests. If we get more bugs like this i’ll consider it.