Disable cookie store in app registry service (#4618)
### Description

We've implemented cookie persistence wrong.
Currently we are persisting the cookie every time we apply a new block.
There are two cases when the service restarts:
1. if the minipoolGen is after the snapshot - we fail to sync since the
sync session runner expects a reset cookie as the first update
2. if the minipoolGen is before the snapshot - we will get a reset
cookie, but we are missing events from the snapshot and back until our
minipoolGen

I am disabling the persistence for now, until we fix those issues

### Checklist

- [ ] Tests added where required
- [ ] Documentation updated where applicable
- [ ] Changes adhere to the repository's contribution guidelines